feat(login)+desing and separation of modal component
